feat(YOLOX): lint code and add a simple workflow (#400)
