+++ exps/example/custom/nano.py
@@ -3,6 +3,7 @@
 # Copyright (c) Megvii, Inc. and its affiliates.
 
 import os
+
 import torch.nn as nn
 
 from yolox.exp import Exp as MyExp
