+++ tools/eval.py
@@ -2,6 +2,10 @@
 # -*- coding:utf-8 -*-
 # Copyright (c) Megvii, Inc. and its affiliates.
 
+import argparse
+import os
+import random
+import warnings
 from loguru import logger
 
 import torch
@@ -10,12 +14,7 @@ from torch.nn.parallel import DistributedDataParallel as DDP
 
 from yolox.core import launch
 from yolox.exp import get_exp
-from yolox.utils import configure_nccl, fuse_model, get_local_rank, get_model_info, setup_logger
-
-import argparse
-import os
-import random
-import warnings
+from yolox.utils import fuse_model, get_model_info, setup_logger
 
 
 def make_parser():
