+++ tools/export_onnx.py
@@ -2,6 +2,8 @@
 # -*- coding:utf-8 -*-
 # Copyright (c) Megvii, Inc. and its affiliates.
 
+import argparse
+import os
 from loguru import logger
 
 import torch
@@ -10,9 +12,6 @@ from torch import nn
 from yolox.exp import get_exp
 from yolox.models.network_blocks import SiLU
 from yolox.utils import replace_module
-
-import argparse
-import os
 
 
 def make_parser():
