+++ yolox/data/data_prefetcher.py
@@ -2,12 +2,12 @@
 # -*- coding:utf-8 -*-
 # Copyright (c) Megvii, Inc. and its affiliates.
 
+import random
+
 import torch
 import torch.distributed as dist
 
 from yolox.utils import synchronize
-
-import random
 
 
 class DataPrefetcher:
