+++ yolox/evaluators/voc_evaluator.py
@@ -2,6 +2,10 @@
 # -*- coding:utf-8 -*-
 # Copyright (c) Megvii, Inc. and its affiliates.
 
+import sys
+import tempfile
+import time
+from collections import ChainMap
 from loguru import logger
 from tqdm import tqdm
 
@@ -10,11 +14,6 @@ import numpy as np
 import torch
 
 from yolox.utils import gather, is_main_process, postprocess, synchronize, time_synchronized
-
-import sys
-import tempfile
-import time
-from collections import ChainMap
 
 
 class VOCEvaluator:
