+++ yolox/utils/checkpoint.py
@@ -1,12 +1,11 @@
 #!/usr/bin/env python3
 # -*- coding:utf-8 -*-
 # Copyright (c) 2014-2021 Megvii Inc. All rights reserved.
+import os
+import shutil
 from loguru import logger
 
 import torch
-
-import os
-import shutil
 
 
 def load_ckpt(model, ckpt):
