+++ yolox/utils/logger.py
@@ -2,11 +2,10 @@
 # -*- coding:utf-8 -*-
 # Copyright (c) 2014-2021 Megvii Inc. All rights reserved.
 
-from loguru import logger
-
 import inspect
 import os
 import sys
+from loguru import logger
 
 
 def get_caller_name(depth=0):
