+++ yolox/utils/metric.py
@@ -1,14 +1,14 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 # Copyright (c) 2014-2021 Megvii Inc. All rights reserved.
-import numpy as np
-
-import torch
-
 import functools
 import os
 import time
 from collections import defaultdict, deque
+
+import numpy as np
+
+import torch
 
 __all__ = [
     "AverageMeter",
